+++ app/types.d.ts
@@ -1,0 +1,10 @@
+export type Breed = {
+    name: string;
+    id: string;
+};
+
+export type CatData = {
+    breeds: Breed[];
+    id: string;
+    url: string;
+};
